// angular/src/app/admin/ecommerce/Models/product-model.ts
export interface ProductModel {
    id?: number;
    productName: string;
    manufacturerName: string;
    manufacturerBrand: string;
    price: number;
    quantity: number;
    category: string;
    features: string;
    description: string;
    metaTitle: string;
    metaKeywords: string;
    metaDescription: string;
    attachmentId:number
  }